fix clang compilation warning
Submitted on behalf of Guanqun.Lu@gmail.com

Review URL: http://codereview.appspot.com/5833045/


git-svn-id: http://skia.googlecode.com/svn/trunk@3472 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ tests/UnicodeTest.cpp
@@ -69,9 +69,9 @@ static void test_textencodings(skiatest::Reporter* reporter) {
     paint.setTextEncoding(SkPaint::kUTF32_TextEncoding);
     int count32 = paint.textToGlyphs(text32, len32, glyphs32);
 
-    REPORTER_ASSERT(reporter, len8 == count8);
-    REPORTER_ASSERT(reporter, len8 == count16);
-    REPORTER_ASSERT(reporter, len8 == count32);
+    REPORTER_ASSERT(reporter, (int)len8 == count8);
+    REPORTER_ASSERT(reporter, (int)len8 == count16);
+    REPORTER_ASSERT(reporter, (int)len8 == count32);
     
     REPORTER_ASSERT(reporter, !memcmp(glyphs8, glyphs16, count8 * sizeof(uint16_t)));
     REPORTER_ASSERT(reporter, !memcmp(glyphs8, glyphs32, count8 * sizeof(uint16_t)));
